backend/requirements.txt
Flask==2.2.2
google-cloud-dialogflow==2.38.0
google-auth==2.38.0
requests==2.32.3
